logic is separated from the UI
