+++ rdpWrapper/Themes/CustomTheme.cs
@@ -22,11 +22,11 @@ namespace sergiye.Common {
     public DarkTheme() : base("dark", "Dark") {
       ForegroundColor = ColorTranslator.FromHtml("#DADADA");
       BackgroundColor = Color.Black;
-      HyperlinkColor = ColorTranslator.FromHtml("#90E6E8");
+      HyperlinkColor = Color.OrangeRed;
       SelectedForegroundColor = ColorTranslator.FromHtml("#DADADA");
       SelectedBackgroundColor = ColorTranslator.FromHtml("#2170CF");
-      LineColor = ColorTranslator.FromHtml("#070A12");
-      StrongLineColor = ColorTranslator.FromHtml("#091217");
+      LineColor = Color.Orange;
+      StrongLineColor = Color.DarkOrange;
       WindowTitlebarFallbackToImmersiveDarkMode = true;
       StatusOkColor = Color.LawnGreen;
       StatusInfoColor = Color.Gold;
